# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI20K Agent System",
    "description": "AI20K Agent System API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/data-sources": {
      "get": {
        "tags": [
          "Data Sources"
        ],
        "summary": "List Data Sources",
        "description": "Liệt kê nguồn nếu người dùng là thành viên dự án.",
        "operationId": "listProjectDataSources",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa nguồn dữ liệu",
              "title": "Project Id"
            },
            "description": "ID dự án chứa nguồn dữ liệu"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DataSourceListResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/data-sources/upload": {
      "post": {
        "tags": [
          "Data Sources"
        ],
        "summary": "Upload Data Sources",
        "description": "Upload và phân tích batch file CSV nếu người dùng là OWNER.",
        "operationId": "uploadProjectDataSources",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa nguồn dữ liệu",
              "title": "Project Id"
            },
            "description": "ID dự án chứa nguồn dữ liệu"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadProjectDataSources"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UploadDataSourcesResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/data-sources/{source_id}/preview": {
      "get": {
        "tags": [
          "Data Sources"
        ],
        "summary": "Get Data Source Preview",
        "description": "Đọc preview CSV theo yêu cầu, không lưu bản sao trong database.",
        "operationId": "getProjectDataSourcePreview",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa nguồn dữ liệu",
              "title": "Project Id"
            },
            "description": "ID dự án chứa nguồn dữ liệu"
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID nguồn dữ liệu",
              "title": "Source Id"
            },
            "description": "ID nguồn dữ liệu"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DataSourcePreviewResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/data-sources/{source_id}/tables/{table_name}/columns/{column_name}": {
      "patch": {
        "tags": [
          "Data Sources"
        ],
        "summary": "Update Data Source Column",
        "description": "Cập nhật một phần metadata của cột nếu người dùng là OWNER.",
        "operationId": "updateProjectDataSourceColumn",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa nguồn dữ liệu",
              "title": "Project Id"
            },
            "description": "ID dự án chứa nguồn dữ liệu"
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID nguồn dữ liệu",
              "title": "Source Id"
            },
            "description": "ID nguồn dữ liệu"
          },
          {
            "name": "table_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Tên bảng",
              "title": "Table Name"
            },
            "description": "Tên bảng"
          },
          {
            "name": "column_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Tên cột",
              "title": "Column Name"
            },
            "description": "Tên cột"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDataSourceColumnRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DataSourceResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/data-sources/{source_id}": {
      "delete": {
        "tags": [
          "Data Sources"
        ],
        "summary": "Delete Data Source",
        "description": "Xóa nguồn và file vật lý nếu người dùng là OWNER.",
        "operationId": "deleteProjectDataSource",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa nguồn dữ liệu",
              "title": "Project Id"
            },
            "description": "ID dự án chứa nguồn dữ liệu"
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID nguồn dữ liệu",
              "title": "Source Id"
            },
            "description": "ID nguồn dữ liệu"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Code",
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông báo lỗi"
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "Mã lỗi ổn định cho client"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Danh sách chi tiết lỗi theo field"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "error_code"
        ],
        "title": "ApiErrorResponse",
        "description": "Error envelope được Global Exception Handler trả về."
      },
...
      "ApiResponse_DataSourceListResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataSourceListResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[DataSourceListResponse]"
      },
      "ApiResponse_DataSourcePreviewResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataSourcePreviewResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[DataSourcePreviewResponse]"
      },
      "ApiResponse_DataSourceResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataSourceResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[DataSourceResponse]"
      },
...
      "ApiResponse_UploadDataSourcesResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UploadDataSourcesResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[UploadDataSourcesResponse]"
      },
...
      "Body_uploadProjectDataSources": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "Tối đa 20 file CSV"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_uploadProjectDataSources"
      },
...
      "CheckConstraintDto": {
        "properties": {
          "type": {
            "type": "string",
            "const": "CHECK",
            "title": "Type",
            "description": "Loại constraint"
          },
          "expression": {
            "type": "string",
            "minLength": 1,
            "title": "Expression",
            "description": "Biểu thức CHECK chính thức"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "expression"
        ],
        "title": "CheckConstraintDto",
        "description": "Constraint kiểm tra."
      },
...
      "ColumnDataType": {
        "type": "string",
        "enum": [
          "TEXT",
          "CATEGORY",
          "INTEGER",
          "NUMBER",
          "DECIMAL",
          "DATE",
          "TIME",
          "DATETIME",
          "BOOLEAN"
        ],
        "title": "ColumnDataType",
        "description": "Kiểu dữ liệu hiển thị của một cột nguồn."
      },
...
      "DataSourceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DataSourceResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "Danh sách nguồn"
          },
          "can_edit": {
            "type": "boolean",
            "title": "Can Edit",
            "description": "Actor có quyền chỉnh sửa"
          }
        },
        "type": "object",
        "required": [
          "items",
          "can_edit"
        ],
        "title": "DataSourceListResponse",
        "description": "Danh sách nguồn kèm quyền chỉnh sửa của actor."
      },
      "DataSourcePreviewResponse": {
        "properties": {
          "rows": {
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "type": "array",
            "title": "Rows",
            "description": "Các dòng preview"
          },
          "total_rows": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Rows",
            "description": "Tổng số dòng"
          }
        },
        "type": "object",
        "required": [
          "rows",
          "total_rows"
        ],
        "title": "DataSourcePreviewResponse",
        "description": "Preview đọc lười từ file nguồn."
      },
...
      "DefaultConstraintDto": {
        "properties": {
          "type": {
            "type": "string",
            "const": "DEFAULT",
            "title": "Type",
            "description": "Loại constraint"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "Giá trị mặc định hoặc biểu thức dạng chuỗi"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "value"
        ],
        "title": "DefaultConstraintDto",
        "description": "Constraint giá trị mặc định."
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Đường dẫn field không hợp lệ"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mô tả lỗi validation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Chi tiết validation của một trường request."
      },
...
      "ForeignKeyConstraintDto": {
        "properties": {
          "type": {
            "type": "string",
            "const": "FOREIGN_KEY",
            "title": "Type",
            "description": "Loại constraint"
          },
          "reference_table": {
            "type": "string",
            "minLength": 1,
            "title": "Reference Table",
            "description": "Bảng được tham chiếu"
          },
          "reference_column": {
            "type": "string",
            "minLength": 1,
            "title": "Reference Column",
            "description": "Cột được tham chiếu"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "reference_table",
          "reference_column"
        ],
        "title": "ForeignKeyConstraintDto",
        "description": "Constraint khóa ngoại."
      },
...
      "UniqueConstraintDto": {
        "properties": {
          "type": {
            "type": "string",
            "const": "UNIQUE",
            "title": "Type",
            "description": "Loại constraint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type"
        ],
        "title": "UniqueConstraintDto",
        "description": "Constraint duy nhất."
      },
...
      "UpdateDataSourceColumnRequest": {
        "properties": {
          "data_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ColumnDataType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Kiểu hiển thị mới"
          },
          "distinct_values": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array",
                "maxItems": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Distinct Values",
            "description": "Tập giá trị phân biệt cần hiển thị"
          },
          "constraints": {
            "anyOf": [
              {
                "items": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ForeignKeyConstraintDto"
                    },
                    {
                      "$ref": "#/components/schemas/UniqueConstraintDto"
                    },
                    {
                      "$ref": "#/components/schemas/CheckConstraintDto"
                    },
                    {
                      "$ref": "#/components/schemas/DefaultConstraintDto"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "CHECK": "#/components/schemas/CheckConstraintDto",
                      "DEFAULT": "#/components/schemas/DefaultConstraintDto",
                      "FOREIGN_KEY": "#/components/schemas/ForeignKeyConstraintDto",
                      "UNIQUE": "#/components/schemas/UniqueConstraintDto"
                    }
                  }
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints",
            "description": "Danh sách constraint chính thức của cột"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UpdateDataSourceColumnRequest",
        "description": "Payload cập nhật một phần metadata cột."
      },
...
      "UploadDataSourcesResponse": {
        "properties": {
          "data_sources": {
            "items": {
              "$ref": "#/components/schemas/DataSourceResponse"
            },
            "type": "array",
            "title": "Data Sources",
            "description": "Nguồn đã lưu"
          },
          "total_files_uploaded": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Files Uploaded",
            "description": "Số file đã tải lên"
          }
        },
        "type": "object",
        "required": [
          "data_sources",
          "total_files_uploaded"
        ],
        "title": "UploadDataSourcesResponse",
        "description": "Kết quả một batch upload CSV."
      },
...
    }
  }
}